% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{appends}
\alias{appends}
\title{appends element to a list object}
\usage{
appends(lists, to_add)
}
\arguments{
\item{lists}{The original list}

\item{to_add}{The element to be added to the list}
}
\value{
A new list with the element added at the end
}
\description{
This function appends new element to an existing list object at the end (the original list can be empty)
}
