% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{batch_projection}
\alias{batch_projection}
\title{batch projection of data in siRNA files to classification boundary}
\usage{
batch_projection(directory, index, center, std, weight)
}
\arguments{
\item{directory}{Directory of the folder that contains all data files to be projected}

\item{index}{A vector of index for features to be used for projection}

\item{center}{Center of typical cells, outputs of SVM()}

\item{std}{Standard deviations of typical cells, outputs of SVM()}

\item{weights}{Weights of each feature (i.e. direction of the classification boundary), outputs of SVM()}
}
\value{
percentages A vector of percentage of cells with positive projection for each data file

cell.counts A vector of cell numbers in each data file

names A vector of names for each data file
}
\description{
This function projects data in multiple data files to the dreiction of classification boundray identified in SVM()
}
