% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{combine_data}
\alias{combine_data}
\title{combine all data files within a certain directory}
\usage{
combine_data(directory, pattern = NULL)
}
\arguments{
\item{directory}{The directory of the folder that stores all the data files to be combined}

\item{pattern}{An optional regular expression. Only files match this expression with be combined. Default value is NULL.}
}
\value{
A data frame that contains all data
}
\description{
This function combines all data files in a given folder
}
