% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{.find_peaks}
\alias{.find_peaks}
\title{Find peaks of a 1D distribution}
\usage{
.find_peaks(x, y = NULL, num_peaks = NULL, adjust = 2,
  plot = FALSE, ...)
}
\description{
This function finds peaks for a 1D density distribution and return positions of these peaks in a list
}
\details{
This function is taken from flowDensity package at ()
}
