% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{hits_identification}
\alias{hits_identification}
\title{Identify important siRNA hits}
\usage{
hits_identification(disease_percent, siRNA_percent, std_siRNA_percent,
  siRNA_number, siRNA_name, number_of_plates)
}
\arguments{
\item{disease_percent}{A vector of healthy-like cell percentage in disease control samples}

\item{siRNA_percent}{A vector of healthy-like cell percentage in siRNA samples}

\item{std_siRNA_percent}{A vector of standard deviations of each siRNA sample across replicate experiments}

\item{siRNA_number}{A vector of ratio between the number of cells in each siRNA sample and mean cell numbers in healthy control samples}

\item{siRNA_name}{A character vector contains the names of each siRNA}

\item{number_of_plates}{The number of replicate experiments}
}
\value{
A character vector containing the names of siRNA hits
}
\description{
This function identifies important siRNA hits as the ones with percentage of healthy-like cells above a threshold
}
