% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{SVM}
\alias{SVM}
\title{performs SVM to classify data points in two matrices}
\usage{
SVM(healthy, disease, feature_index = c(1:ncol(healthy)))
}
\arguments{
\item{healthy}{Matrix of healthy cells to be classified}

\item{disease}{Matrix of diseased cells to be classified}

\item{feature_index}{A vector contain index of measurements used in classification, should be a vector of integers, length must be larger than 1}
}
\value{
accuracy Accuracy of the classification

weightnorm Normalized weights of each feature. Negative weights are higher in disease matrix

center The center (average) of all data points (healthy and disease combined). Can be used to normalize test datasets

std The standard deviation of all data points (healthy and disease combined). Can be used to normalize test datasets

SVM_bn adjusted constant for calculating distance between data point to classification boundary in test data points
}
\description{
This function applies SVM in kernlab package to classify two groups of data points stored in two matrices and returns the accuracy 
of SVM classification, and the direction of classification boundary
}
