% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{SVM_cv}
\alias{SVM_cv}
\title{performs SVM classification and testing}
\usage{
SVM_cv(controls, exper, feature_index, test.control, test.exp)
}
\arguments{
\item{controls}{Training matrix of healthy cells to be classified}

\item{exper}{Training matrix of diseased cells to be classified}

\item{feature_index}{A vector contain index of measurements used in classification, should be a vector of integers, length must be larger than 1}

\item{test.control}{Testing matrix of healthy cells}

\item{test.exp}{Testing matrix of diseased cells}
}
\value{
accuracy Classificaiton accuracy of the training dataset

test Classification accuracy of the test dataset

weight Normalized weights of the features
}
\description{
This function applies SVM in kernlab package to classify two groups of data points stored in two matrices, 
applies its classification boundary to testing dataset and outputs corresponding accuracy.
}
