% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCells_functions.R
\name{typical_cell_selection}
\alias{typical_cell_selection}
\title{selects typical cells}
\usage{
typical_cell_selection(input, number, index, method = "peak")
}
\arguments{
\item{input}{Input matrix that contains all cells}

\item{number}{The number of typical cells to select}

\item{index}{Indices of measurements based on which typical cells are selected, should be a vector of integers}

\item{method}{Methods to define the center of typical cells, can choose from "mean" and "peak"}
}
\value{
A list contains two elements, the first element is a matrix contain typical cells, and the second element is a vector of index for typical cells
}
\description{
This function selects typical cells and returns selected typical cells as a matrix
}
